Elimina manager de entidad innecesaria PuntosManager
