+++ app/src/main/java/com/ssanchez/puntosyrecompensas/UsuariosManager.kt
@@ -1,5 +1,6 @@
 package com.ssanchez.puntosyrecompensas
 
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseReference
@@ -26,6 +27,33 @@ object UsuarioManager {
                 onResult(null) // Devuelve null en caso de error
             }
         })
+    }
+
+    fun getPuntos(mauth: FirebaseAuth, callback: (Int) -> Unit) {
+        val puntosRef = usuariosRef.child(mauth.currentUser!!.uid).child("saldoPuntos")
+
+        puntosRef.addListenerForSingleValueEvent(object : ValueEventListener {
+            override fun onDataChange(snapshot: DataSnapshot) {
+                val saldoPuntos = snapshot.getValue(Int::class.java) ?: 0
+                callback(saldoPuntos)  // Llamada al callback con saldoPuntos
+            }
+
+            override fun onCancelled(error: DatabaseError) {
+                callback(0) // En caso de error, devuelve 0 como valor predeterminado
+            }
+        })
+    }
+
+    fun setPuntos(mauth: FirebaseAuth, puntos: Int, callback: (Boolean) -> Unit) {
+        val puntosRef = usuariosRef.child(mauth.currentUser!!.uid).child("saldoPuntos")
+
+        puntosRef.setValue(puntos).addOnCompleteListener { task ->
+            if (task.isSuccessful) {
+                callback(true) // Éxito
+            } else {
+                callback(false) // Error
+            }
+        }
     }
 
 }
